+++ data-raw/preprocess_data.R
@@ -1090,6 +1090,16 @@ parameters_Huang2013_hd <- readr::read_csv(
     Subregion = stringr::str_squish(as.character(Subregion))
   )
 
+parameters_SharmaParton2007_hd <- readr::read_csv(
+  "data-raw/parameters_SharmaParton2007_hd.csv",
+  show_col_types = FALSE
+) %>%
+  mutate(
+    Species = standardize_species_code(Species, keep_all = FALSE),
+    fit = stringr::str_squish(as.character(fit)),
+    AIC = as.numeric(AIC)
+  )
+
 # combine all into one ####
 
 internal_objs <- c(
@@ -1118,6 +1128,7 @@ internal_objs <- c(
   "parameters_Thrower1994",
   "parameters_Huang1994_si",
   "parameters_Huang2013_hd",
+  "parameters_SharmaParton2007_hd",
   "parameters_Carmean1989",
   "parameters_Carmean1996",
   "parameters_QC_IQS2013",